Refs #21266 Change namespaces, README  and regenerate types
