+++ examples/CMakeLists.txt
@@ -19,11 +19,8 @@ add_subdirectory(cpp/content_filter)
 add_subdirectory(cpp/custom_payload_pool)
 add_subdirectory(cpp/dds)
 add_subdirectory(cpp/delivery_mechanisms)
-<<<<<<< HEAD
 add_subdirectory(cpp/discovery_server)
-=======
 add_subdirectory(cpp/dds_security)
->>>>>>> 420ed93df (Refs #21266 Add secured hello world example)
 add_subdirectory(cpp/hello_world)
 add_subdirectory(cpp/rtps)
 add_subdirectory(cpp/xtypes)
