+++ examples/cpp/security/Application.cpp
@@ -20,19 +20,19 @@
 #include "Application.hpp"
 
 #include "CLIParser.hpp"
-#include "SubscriberApp.hpp"
 #include "PublisherApp.hpp"
+#include "SubscriberApp.hpp"
 
 using namespace eprosima::fastdds::dds;
 
 namespace eprosima {
 namespace fastdds {
 namespace examples {
-namespace hello_world {
+namespace security {
 
 //! Factory method to create a publisher or subscriber
 std::shared_ptr<Application> Application::make_app(
-        const CLIParser::hello_world_config& config,
+        const CLIParser::security_config& config,
         const std::string& topic_name)
 {
     std::shared_ptr<Application> entity;
@@ -52,7 +52,7 @@ std::shared_ptr<Application> Application::make_app(
     return entity;
 }
 
-} // namespace hello_world
+} // namespace security
 } // namespace examples
 } // namespace fastdds
 } // namespace eprosima
