+++ examples/cpp/security/HelloWorld.hpp
@@ -19,8 +19,8 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef _FAST_DDS_GENERATED_HELLOWORLD_HPP_
-#define _FAST_DDS_GENERATED_HELLOWORLD_HPP_
+#ifndef FAST_DDS_GENERATED__HELLOWORLD_HPP
+#define FAST_DDS_GENERATED__HELLOWORLD_HPP
 
 #include <cstdint>
 #include <string>
