+++ examples/cpp/security/HelloWorldCdrAux.hpp
@@ -19,8 +19,8 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_HPP_
-#define _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_HPP_
+#ifndef FAST_DDS_GENERATED__HELLOWORLDCDRAUX_HPP
+#define FAST_DDS_GENERATED__HELLOWORLDCDRAUX_HPP
 
 #include "HelloWorld.hpp"
 
@@ -42,5 +42,5 @@ eProsima_user_DllExport void serialize_key(
 } // namespace fastcdr
 } // namespace eprosima
 
-#endif // _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_HPP_
+#endif // FAST_DDS_GENERATED__HELLOWORLDCDRAUX_HPP
