+++ examples/cpp/security/HelloWorldCdrAux.ipp
@@ -19,8 +19,8 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_IPP_
-#define _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_IPP_
+#ifndef FAST_DDS_GENERATED__HELLOWORLDCDRAUX_IPP
+#define FAST_DDS_GENERATED__HELLOWORLDCDRAUX_IPP
 
 #include "HelloWorldCdrAux.hpp"
 
@@ -122,5 +122,5 @@ void serialize_key(
 } // namespace fastcdr
 } // namespace eprosima
 
-#endif // _FAST_DDS_GENERATED_HELLOWORLDCDRAUX_IPP_
+#endif // FAST_DDS_GENERATED__HELLOWORLDCDRAUX_IPP
