+++ examples/cpp/security/HelloWorldPubSubTypes.cxx
@@ -19,7 +19,7 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#include "HelloWorldPubSubTypes.h"
+#include "HelloWorldPubSubTypes.hpp"
 
 #include <fastdds/dds/log/Log.hpp>
 #include <fastdds/rtps/common/CdrSerialization.hpp>
