+++ examples/cpp/security/HelloWorldPubSubTypes.hpp
@@ -13,21 +13,21 @@
 // limitations under the License.
 
 /*!
- * @file HelloWorldPubSubTypes.h
+ * @file HelloWorldPubSubTypes.hpp
  * This header file contains the declaration of the serialization functions.
  *
  * This file was generated by the tool fastddsgen.
  */
 
 
-#ifndef _FAST_DDS_GENERATED_HELLOWORLD_PUBSUBTYPES_H_
-#define _FAST_DDS_GENERATED_HELLOWORLD_PUBSUBTYPES_H_
+#ifndef FAST_DDS_GENERATED__HELLOWORLD_PUBSUBTYPES_HPP
+#define FAST_DDS_GENERATED__HELLOWORLD_PUBSUBTYPES_HPP
 
 #include <fastdds/dds/core/policy/QosPolicies.hpp>
 #include <fastdds/dds/topic/TopicDataType.hpp>
-#include <fastdds/rtps/common/InstanceHandle.h>
-#include <fastdds/rtps/common/SerializedPayload.h>
-#include <fastdds/utils/md5.h>
+#include <fastdds/rtps/common/InstanceHandle.hpp>
+#include <fastdds/rtps/common/SerializedPayload.hpp>
+#include <fastdds/utils/md5.hpp>
 
 #include "HelloWorld.hpp"
 
@@ -106,7 +106,7 @@ public:
     }
 
     eProsima_user_DllExport inline bool is_plain(
-        eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
+            eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
     {
         static_cast<void>(data_representation);
         return false;
@@ -124,10 +124,10 @@ public:
 
 #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
 
-    MD5 m_md5;
+    eprosima::fastdds::MD5 m_md5;
     unsigned char* m_keyBuffer;
 
 };
 
-#endif // _FAST_DDS_GENERATED_HELLOWORLD_PUBSUBTYPES_H_
+#endif // FAST_DDS_GENERATED__HELLOWORLD_PUBSUBTYPES_HPP
