+++ examples/cpp/security/HelloWorldTypeObjectSupport.cxx
