+++ examples/cpp/security/HelloWorldTypeObjectSupport.hpp
@@ -19,8 +19,8 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef _FAST_DDS_GENERATED_HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP_
-#define _FAST_DDS_GENERATED_HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP_
+#ifndef FAST_DDS_GENERATED__HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP
+#define FAST_DDS_GENERATED__HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP
 
 #include <fastdds/dds/xtypes/type_representation/TypeObject.hpp>
 
@@ -53,4 +53,4 @@ eProsima_user_DllExport void register_HelloWorld_type_identifier(
 
 #endif // DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
-#endif // _FAST_DDS_GENERATED_HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP_
+#endif // FAST_DDS_GENERATED__HELLOWORLD_TYPE_OBJECT_SUPPORT_HPP
