+++ examples/cpp/security/PublisherApp.hpp
@@ -13,53 +13,49 @@
 // limitations under the License.
 
 /**
- * @file SubscriberApp.hpp
+ * @file PublisherApp.hpp
  *
  */
 
-#ifndef _FASTDDS_HELLO_WORLD_LISTENER_SUBSCRIBER_APP_HPP_
-#define _FASTDDS_HELLO_WORLD_LISTENER_SUBSCRIBER_APP_HPP_
+#ifndef FASTDDS_EXAMPLES_CPP_SECURITY__HELLO_WORLD_PUBLISHER_APP_HPP
+#define FASTDDS_EXAMPLES_CPP_SECURITY__HELLO_WORLD_PUBLISHER_APP_HPP
 
 #include <condition_variable>
 
 #include <fastdds/dds/domain/DomainParticipant.hpp>
-#include <fastdds/dds/subscriber/DataReaderListener.hpp>
+#include <fastdds/dds/publisher/DataWriterListener.hpp>
 #include <fastdds/dds/topic/TypeSupport.hpp>
 
-#include "CLIParser.hpp"
-#include "HelloWorldPubSubTypes.h"
 #include "Application.hpp"
+#include "CLIParser.hpp"
+#include "HelloWorld.hpp"
 
 using namespace eprosima::fastdds::dds;
 
 namespace eprosima {
 namespace fastdds {
 namespace examples {
-namespace hello_world {
+namespace security {
 
-class SubscriberApp : public Application,  public DataReaderListener
+class PublisherApp : public Application, public DataWriterListener
 {
 public:
 
-    SubscriberApp(
+    PublisherApp(
             const CLIParser::entity_config& config,
             const std::string& topic_name);
 
-    ~SubscriberApp();
+    ~PublisherApp();
 
-    //! Subscription callback
-    void on_data_available(
-            DataReader* reader) override;
+    //! Publisher matched method
+    void on_publication_matched(
+            DataWriter* writer,
+            const PublicationMatchedStatus& info) override;
 
-    //! Subscriber matched method
-    void on_subscription_matched(
-            DataReader* reader,
-            const SubscriptionMatchedStatus& info) override;
-
-    //! Run subscriber
+    //! Run publisher
     void run() override;
 
-    //! Trigger the end of execution
+    //! Stop publisher
     void stop() override;
 
 private:
@@ -67,32 +63,38 @@ private:
     //! Return the current state of execution
     bool is_stopped();
 
+    //! Publish a sample
+    bool publish();
+
     HelloWorld hello_;
 
     DomainParticipant* participant_;
 
-    Subscriber* subscriber_;
+    Publisher* publisher_;
 
     Topic* topic_;
 
-    DataReader* reader_;
+    DataWriter* writer_;
 
     TypeSupport type_;
 
+    const uint32_t period_ms_; // sleep time (ms) between consecutive samples
+
+    int16_t matched_;
+
     uint16_t samples_;
 
-    uint16_t received_samples_;
+    std::mutex mutex_;
+
+    std::condition_variable cv_;
 
     std::atomic<bool> stop_;
 
-    mutable std::mutex terminate_cv_mtx_;
-
-    std::condition_variable terminate_cv_;
 };
 
-} // namespace hello_world
+} // namespace security
 } // namespace examples
 } // namespace fastdds
 } // namespace eprosima
 
-#endif /* _FASTDDS_HELLO_WORLD_LISTENER_SUBSCRIBER_APP_HPP_ */
+#endif /* FASTDDS_EXAMPLES_CPP_SECURITY__HELLO_WORLD_PUBLISHER_APP_HPP */
