+++ examples/cpp/security/certs/governance.smime
