+++ examples/cpp/security/certs/mainpubcert.pem
