+++ examples/cpp/security/certs/mainpubkey.pem
