+++ examples/cpp/security/certs/permissions.smime
