+++ examples/cpp/security/hello_world_profile.xml
